+++ pollbot/views/release.py
@@ -5,7 +5,7 @@ from pollbot import PRODUCTS
 from ..exceptions import TaskError
 from ..tasks.archives import archives, archives_date, archives_date_l10n
 from ..tasks.bedrock import release_notes, security_advisories, download_links, get_releases
-from ..tasks.product_details import product_details
+from ..tasks.product_details import product_details, devedition_and_beta_in_sync
 from ..utils import Channel, get_version_channel
 
 
@@ -38,6 +38,7 @@ bedrock_release_notes = status_response(release_notes)
 bedrock_security_advisories = status_response(security_advisories)
 bedrock_download_links = status_response(download_links)
 product_details = status_response(product_details)
+devedition_beta_check = status_response(devedition_and_beta_in_sync)
 
 
 async def view_get_releases(request):
@@ -62,6 +63,7 @@ CHECKS_TITLE = {
     "security-advisories": "Security advisories",
     "download-links": "Download links",
     "product-details": "Product details",
+    "devedition-beta-matches": "Devedition and Beta versions matches",
 }
 
 
@@ -73,7 +75,8 @@ CHECKS = OrderedDict(
         "release-notes": [Channel.ESR, Channel.RELEASE, Channel.BETA, Channel.NIGHTLY],
         "security-advisories": [Channel.ESR, Channel.RELEASE],
         "download-links": [Channel.ESR, Channel.RELEASE, Channel.BETA, Channel.NIGHTLY],
-        "product-details": [Channel.ESR, Channel.RELEASE, Channel.BETA, Channel.NIGHTLY]
+        "product-details": [Channel.ESR, Channel.RELEASE, Channel.BETA, Channel.NIGHTLY],
+        "devedition-beta-matches": [Channel.BETA],
     }.items(), key=lambda t: t[0]))